10월 25일 에러수정

+++ pinguMarket/src/main/java/com/example/demo/User/UserRepository.java
@@ -7,13 +7,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.demo.Entity.User;
 
 
-//public interface UserRepository extends JpaRepository<User, Integer>{
-//	Optional<User> findByUserId(String userId);
-//}
-
-public interface UserRepository extends JpaRepository<User, Long>{
-	// 추후 삭제 필요
+public interface UserRepository extends JpaRepository<User, Integer>{
 	Optional<User> findById(String username);
 	Optional<User> findByUserId(Integer userId);
 }
+
